Hopefully fix typia problem in e2e tests
