+++ src/app/(i18n)/(authenticated)/(user)/user-preferences/_components/UserPreferencesForm/index.tsx
@@ -18,7 +18,7 @@ import SaveIcon from '@mui/icons-material/SaveRounded';
 import moment from 'moment-timezone';
 import { useSnackbar } from 'notistack';
 
-import type { UserPreferences } from '@backend/users';
+import type { UserPreferences } from '@backend/user-preferences';
 
 import {
     getUserPreferences,
